feat: criar funcionalidade de editar

+++ src/app/pages/clientes/lista-clientes/lista-clientes.component.ts
@@ -15,6 +15,10 @@ import {
 } from '@angular/forms';
 import { FormularioListaClientes } from './formulario-lista-clientes';
 import { Subscription } from 'rxjs';
+import { ActivatedRoute, Router, RouterModule } from '@angular/router';
+import { MomentPipe } from '../../../shared/pipes/moment/moment.pipe';
+import { CelularPipe } from '../../../shared/pipes/celular/celular.pipe';
+import { CepPipe } from '../../../shared/pipes/cep/cep.pipe';
 
 @Component({
   selector: 'app-lista-clientes',
@@ -25,6 +29,10 @@ import { Subscription } from 'rxjs';
     NgbAlertModule,
     FormsModule,
     ReactiveFormsModule,
+    RouterModule,
+    MomentPipe,
+    CelularPipe,
+    CepPipe
   ],
   providers: [NgbProgressbarConfig],
   templateUrl: './lista-clientes.component.html',
@@ -44,7 +52,9 @@ export class ListaClientesComponent implements OnInit, OnDestroy {
   constructor(
     private clientesService: ClientesService,
     private fb: FormBuilder,
-    private progressBarConfig: NgbProgressbarConfig
+    private progressBarConfig: NgbProgressbarConfig,
+    private router: Router,
+    private route: ActivatedRoute
   ) {
     this.configurarProgressBar();
 
@@ -94,13 +104,20 @@ export class ListaClientesComponent implements OnInit, OnDestroy {
   removerCliente(cliente: Cliente) {
     const index = this.clientes?.indexOf(cliente);
 
-    if (index != null) {
+    if (index != null && cliente.id) {
       this.clientes?.splice(index, 1);
       this.clientesService.removerCliente(cliente.id);
       this.clientesFiltrados
         ? this.filtrarCliente(this.form.controls.cliente.value)
         : null;
     }
+  }
+
+  editarCliente(cliente: Cliente) {
+    this.router.navigate(['editar'], {
+      relativeTo: this.route,
+      queryParams: { id: cliente.id },
+    });
   }
 
   private configurarProgressBar() {
